refactor: combine XWindowsClipboards tests into one unittest
